Refactor: notifiable concern
