+++ app/models/job_application.rb
@@ -2,6 +2,7 @@
 
 # Candidacy to a job offer
 class JobApplication < ApplicationRecord
+  include Notifiable
   include Readable
   include Preselectable
   include Rejectable
@@ -56,11 +57,6 @@ class JobApplication < ApplicationRecord
 
   before_validation :set_employer
   before_save :compute_notifications_counter
-  after_update :notify_applicant_new_state, if: -> { new_state_requires_applicant_notification? }
-  after_update :notify_hr_managers_contract_drafting, if: -> { saved_change_to_state? && contract_drafting? }
-  after_update :notify_hr_managers_contract_feedback_waiting, if: -> { saved_change_to_state? && contract_feedback_waiting? }
-  after_update :notify_payroll_managers_contract_received, if: -> { saved_change_to_state? && contract_received? }
-  after_update :notify_managers_affected, if: -> { saved_change_to_state? && affected? }
 
   FINISHED_STATES = %w[affected].freeze
   PROCESSING_STATES = %w[initial phone_meeting to_be_met financial_estimate].freeze
@@ -354,52 +350,6 @@ class JobApplication < ApplicationRecord
   def cant_be_accepted_twice = errors.add(:state, :cant_be_accepted_twice)
 
   def has_accepted_other_job_application? = user.job_applications.where(state: "accepted").where.not(id: id).empty?
-
-  def notify_applicant_new_state
-    ApplicantNotificationsMailer.with(user:, job_offer:, state:).notify_new_state.deliver_later
-  end
-
-  def new_state_requires_applicant_notification? = saved_change_to_state? && NOTIFICATION_STATES.include?(state.to_s)
-
-  def notify_hr_managers_contract_drafting
-    job_offer_hr_managers.each do |administrator|
-      notify_hr_manager_contract_drafting(administrator)
-    end
-  end
-
-  def notify_hr_manager_contract_drafting(administrator)
-    NotificationsMailer.with(administrator:, job_application: self).contract_drafting.deliver_later
-  end
-
-  def notify_hr_managers_contract_feedback_waiting
-    job_offer_hr_managers.each do |administrator|
-      notify_hr_manager_contract_feedback_waiting(administrator)
-    end
-  end
-
-  def notify_hr_manager_contract_feedback_waiting(administrator)
-    NotificationsMailer.with(administrator:, job_application: self).contract_feedback_waiting.deliver_later
-  end
-
-  def notify_payroll_managers_contract_received
-    job_offer_payroll_managers.each do |administrator|
-      notify_payroll_manager_contract_received(administrator)
-    end
-  end
-
-  def notify_payroll_manager_contract_received(administrator)
-    NotificationsMailer.with(administrator:, job_application: self).contract_received.deliver_later
-  end
-
-  def notify_managers_affected
-    (job_offer_hr_managers + job_offer_payroll_managers).each do |administrator|
-      notify_manager_affected(administrator)
-    end
-  end
-
-  def notify_manager_affected(administrator)
-    NotificationsMailer.with(administrator:, job_application: self).affected.deliver_later
-  end
 
   class << self
     def processing_states
